requirements.txt
alembic==1.14.0
amqp==5.3.1
APScheduler==3.11.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
kombu==5.4.2
Mako==1.3.6
prompt_toolkit==3.0.48
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
